Nice! It all inserts into MySql!

+++ newegg/mysqlLoader.py
@@ -25,16 +25,15 @@ source = 'newegg'
 
 sourceFile = dir + source + ext
 
-with conn.cursor() as cursor:
-    query = f'DELETE FROM {table_name}'
-    cursor.execute(query)
-conn.commit()
+purgeTable = True
+if purgeTable:
 
-with conn.cursor() as cursor:
-    query = f'ALTER TABLE {table_name} AUTO_INCREMENT=1'
-    cursor.execute(query)
-conn.commit()
-
+    with conn.cursor() as cursor:
+        query = f'DELETE FROM {table_name}'
+        cursor.execute(query)
+        query = f'ALTER TABLE {table_name} AUTO_INCREMENT=1'
+        cursor.execute(query)
+    conn.commit()
 
 # Open the file and read JSON objects line by line
 with open(sourceFile, "r") as file:
